Refactor provisioners for better modularity
This refactors the provisioning logic so that every provisioner is
defined by its own class. This better encapsulates provisoner logic and
makes adding new provisioners easier.

While this is supposed to be a pure refactoring, it contains a small
functional change, so that provisioning scripts can now not just be
put as string only in the provisioner config, but following a "script"
key as well, similar to all other provisioners.
